pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ali-scraper"
version = "1.0.0"
description = "Alibaba product scraper (KuKirin store) with Playwright stealth"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["scraper", "alibaba", "kukirin", "products", "mongodb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "playwright>=1.40.0",
    "httpx>=0.25.0",
    "pymongo>=4.6.0",
    "dnspython>=2.4.0",
    "certifi>=2023.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
ali-scraper = "ali_scraper.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
